#59 Case Symptoms web ui: * readonly/invisible fields * temperature dropdown * switch temperature to float

+++ sormas-api/src/main/java/de/symeda/sormas/api/symptoms/SymptomsDto.java
@@ -52,7 +52,7 @@ public class SymptomsDto extends DataTransferObject {
 
 	private Date onsetDate;
 	
-	private int temperature;
+	private float temperature;
 	private TemperatureSource temperatureSource;
 	
 	private SymptomState fever;
@@ -99,10 +99,10 @@ public class SymptomsDto extends DataTransferObject {
 	public void setOnsetDate(Date onsetDate) {
 		this.onsetDate = onsetDate;
 	}
-	public int getTemperature() {
+	public float getTemperature() {
 		return temperature;
 	}
-	public void setTemperature(int temperature) {
+	public void setTemperature(float temperature) {
 		this.temperature = temperature;
 	}
 	public TemperatureSource getTemperatureSource() {
